Introduce disabled option flags packing on the server side
Instead of storing disabled option flags separately, they are now packed
into a single integer. This enables to introduce the disable invasion flag
while allowing for easy addition of future additional flags.

This changes the database format. A script is included to help migrate an
existing database to the new format, and the database creation script is
also updated