Adapt server and client to send disabled option flags in packed form
For OPEN_GAME messages and STATUS messages, make the server send out the
information in a format adapted to the client version. For new client, send
the packed information. Make the (ai_)client able to read the packed form.

Also change the way the options are displayed in the client to avoid the
excess of useless columns in the UI.